runtime: decompose weave loading seams
- add nested WEAVE_TIMING phases for candidate loading and ResourcePage collection
- assert timing detail and cache counters in the CLI timing e2e test
- extract runtime errors, mesh-state loading, planning overlay/cache, artifact loaders, candidate loading, request normalization, and timing helpers from the weave façade
- keep public weave runtime exports stable
- update the runtime codebase overview
