+++ src/runtime/weave/errors.ts
@@ -1,0 +1,6 @@
+export class WeaveRuntimeError extends Error {
+  constructor(message: string) {
+    super(message);
+    this.name = "WeaveRuntimeError";
+  }
+}
